requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 25.7.0
_python_abi3_support=1.0=hd8ed1ab_2
anyio=4.10.0=pyhe01879c_0
appnope=0.1.4=pyhd8ed1ab_1
argon2-cffi=25.1.0=pyhd8ed1ab_0
argon2-cffi-bindings=25.1.0=py310h7bdd564_0
arrow=1.3.0=pyhd8ed1ab_1
asttokens=3.0.0=pyhd8ed1ab_1
async-lru=2.0.5=pyh29332c3_0
attrs=25.3.0=pyh71513ae_0
babel=2.17.0=pyhd8ed1ab_0
beautifulsoup4=4.13.5=pyha770c72_0
bleach=6.2.0=pyh29332c3_4
bleach-with-css=6.2.0=h82add2a_4
brotli=1.1.0=h6caf38d_4
brotli-bin=1.1.0=h6caf38d_4
brotli-python=1.1.0=py310h1af2607_4
bzip2=1.0.8=h99b78c6_7
ca-certificates=2025.8.3=hbd8a1cb_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2025.8.3=pyhd8ed1ab_0
cffi=1.17.1=py310h8e7cc95_1
charset-normalizer=3.4.3=pyhd8ed1ab_0
comm=0.2.3=pyhe01879c_0
contourpy=1.3.2=py310h7f4e7e6_0
cpython=3.10.18=py310hd8ed1ab_0
cycler=0.12.1=pyhd8ed1ab_1
debugpy=1.8.16=py310hf6ed2a0_1
decorator=5.2.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_1
exceptiongroup=1.3.0=pyhd8ed1ab_0
executing=2.2.1=pyhd8ed1ab_0
fonttools=4.59.2=py310h5f69134_0
fqdn=1.5.1=pyhd8ed1ab_1
freetype=2.13.3=hce30654_1
h11=0.16.0=pyhd8ed1ab_0
h2=4.3.0=pyhcf101f3_0
hpack=4.1.0=pyhd8ed1ab_0
httpcore=1.0.9=pyh29332c3_0
httpx=0.28.1=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
icu=75.1=hfee45f7_0
idna=3.10=pyhd8ed1ab_1
importlib-metadata=8.7.0=pyhe01879c_1
ipykernel=6.29.5=pyh57ce528_0
ipython=8.37.0=pyh8f84b5b_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
isoduration=20.11.0=pyhd8ed1ab_1
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pyhd8ed1ab_0
joblib=1.5.2=pyhd8ed1ab_0
json5=0.12.1=pyhd8ed1ab_0
jsonpointer=3.0.0=py310hbe9552e_2
jsonschema=4.25.1=pyhe01879c_0
jsonschema-specifications=2025.4.1=pyh29332c3_0
jsonschema-with-format-nongpl=4.25.1=he01879c_0
jupyter-lsp=2.3.0=pyhcf101f3_0
jupyter_client=7.4.9=pyhd8ed1ab_0
jupyter_core=5.8.1=pyh31011fe_0
jupyter_events=0.12.0=pyh29332c3_0
jupyter_server=2.17.0=pyhcf101f3_0
jupyter_server_terminals=0.5.3=pyhd8ed1ab_1
jupyterlab=4.4.6=pyhd8ed1ab_0
jupyterlab_pygments=0.3.0=pyhd8ed1ab_2
jupyterlab_server=2.27.3=pyhd8ed1ab_1
kiwisolver=1.4.9=py310h92dc006_1
krb5=1.21.3=h237132a_0
lark=1.2.2=pyhd8ed1ab_1
lcms2=2.17=h7eeda09_0
lerc=4.0.0=hd64df32_1
libblas=3.9.0=34_h10e41b3_openblas
libbrotlicommon=1.1.0=h6caf38d_4
libbrotlidec=1.1.0=h6caf38d_4
libbrotlienc=1.1.0=h6caf38d_4
libcblas=3.9.0=34_hb3479ef_openblas
libcxx=21.1.0=hf598326_1
libdeflate=1.24=h5773f1b_0
libedit=3.1.20250104=pl5321hafb1f1b_0
libexpat=2.7.1=hec049ff_0
libffi=3.4.6=h1da3d7d_1
libfreetype=2.13.3=hce30654_1
libfreetype6=2.13.3=h1d14073_1
libgfortran=15.1.0=hfdf1602_1
libgfortran5=15.1.0=hb74de2c_1
libjpeg-turbo=3.1.0=h5505292_0
liblapack=3.9.0=34_hc9a63f6_openblas
liblzma=5.8.1=h39f12f2_2
libopenblas=0.3.30=openmp_h60d53f8_2
libpng=1.6.50=h280e0eb_1
libsodium=1.0.18=h27ca646_1
libsqlite=3.50.4=h4237e3c_0
libtiff=4.7.0=h025e3ab_6
libwebp-base=1.6.0=h07db88b_0
libxcb=1.17.0=hdb1d25a_0
libxgboost=3.0.4=cpu_hd76abf7_0
libzlib=1.3.1=h8359307_2
llvm-openmp=21.1.0=hbb9b287_0
markupsafe=3.0.2=py310hc74094e_1
matplotlib=3.10.6=py310hb6292c7_0
matplotlib-base=3.10.6=py310he861c53_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mistune=3.1.4=pyhcf101f3_0
munkres=1.1.4=pyhd8ed1ab_1
nbclient=0.10.2=pyhd8ed1ab_0
nbconvert-core=7.16.6=pyh29332c3_0
nbformat=5.10.4=pyhd8ed1ab_1
ncurses=6.5=h5e97a16_3
nest-asyncio=1.6.0=pyhd8ed1ab_1
notebook=7.4.5=pyhd8ed1ab_0
notebook-shim=0.2.4=pyhd8ed1ab_1
numpy=2.2.6=py310h4d83441_0
openjpeg=2.5.3=h889cd5d_1
openssl=3.5.2=he92f556_0
overrides=7.7.0=pyhd8ed1ab_1
packaging=25.0=pyh29332c3_1
pandas=2.3.2=py310h03dc5a2_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.5=pyhcf101f3_0
patsy=1.0.1=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.3.0=py310h61efb56_0
pip=25.2=pyh8b19718_0
platformdirs=4.4.0=pyhcf101f3_0
prometheus_client=0.22.1=pyhd8ed1ab_0
prompt-toolkit=3.0.52=pyha770c72_0
psutil=7.0.0=py310h7bdd564_1
pthread-stubs=0.4=hd74edd7_1002
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
py-xgboost=3.0.4=cpu_pyh718b53a_0
pycparser=2.22=pyh29332c3_1
pygments=2.19.2=pyhd8ed1ab_0
pyobjc-core=11.1=py310h3532944_1
pyobjc-framework-cocoa=11.1=py310h068f9c9_1
pyparsing=3.2.3=pyhe01879c_2
pysocks=1.7.1=pyha55dd90_7
python=3.10.18=h6cefb37_0_cpython
python-dateutil=2.9.0.post0=pyhe01879c_2
python-fastjsonschema=2.21.2=pyhe01879c_0
python-gil=3.10.18=hd8ed1ab_0
python-json-logger=2.0.7=pyhd8ed1ab_0
python-tzdata=2025.2=pyhd8ed1ab_0
python_abi=3.10=8_cp310
pytz=2025.2=pyhd8ed1ab_0
pyyaml=6.0.2=py310hc74094e_2
pyzmq=24.0.1=py310hc407298_1
qhull=2020.2=h420ef59_5
readline=8.2=h1d1bf99_2
referencing=0.36.2=pyh29332c3_0
requests=2.32.5=pyhd8ed1ab_0
rfc3339-validator=0.1.4=pyhd8ed1ab_1
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rfc3987-syntax=1.1.0=pyhe01879c_1
rpds-py=0.27.1=py310h7018d9b_1
scikit-learn=1.7.1=py310hce580d2_0
scipy=1.15.2=py310h32ab4ed_0
seaborn=0.13.2=hd8ed1ab_3
seaborn-base=0.13.2=pyhd8ed1ab_3
send2trash=1.8.3=pyh31c8845_1
setuptools=80.9.0=pyhff2d567_0
six=1.17.0=pyhe01879c_1
sniffio=1.3.1=pyhd8ed1ab_1
soupsieve=2.8=pyhd8ed1ab_0
stack_data=0.6.3=pyhd8ed1ab_1
statsmodels=0.14.5=py310h307809b_0
terminado=0.18.1=pyh31c8845_0
threadpoolctl=3.6.0=pyhecae5ae_0
tinycss2=1.4.0=pyhd8ed1ab_0
tk=8.6.13=h892fb3f_2
tomli=2.2.1=pyhe01879c_2
tornado=6.5.2=py310h7bdd564_0
traitlets=5.14.3=pyhd8ed1ab_1
types-python-dateutil=2.9.0.20250822=pyhd8ed1ab_0
typing-extensions=4.15.0=h396c80c_0
typing_extensions=4.15.0=pyhcf101f3_0
typing_utils=0.1.0=pyhd8ed1ab_1
tzdata=2025b=h78e105d_0
unicodedata2=16.0.0=py310h7bdd564_1
uri-template=1.3.0=pyhd8ed1ab_1
urllib3=2.5.0=pyhd8ed1ab_0
wcwidth=0.2.13=pyhd8ed1ab_1
webcolors=24.11.1=pyhd8ed1ab_0
webencodings=0.5.1=pyhd8ed1ab_3
websocket-client=1.8.0=pyhd8ed1ab_1
wheel=0.45.1=pyhd8ed1ab_1
xorg-libxau=1.0.12=h5505292_0
xorg-libxdmcp=1.1.5=hd74edd7_0
yaml=0.2.5=h925e9cb_3
zeromq=4.3.5=hcc0f68c_4
zipp=3.23.0=pyhd8ed1ab_0
zstandard=0.24.0=py310hc131fc1_0
zstd=1.5.7=h6491c7d_2
